Ch 4 Checkpoint | Move helper functions below struct
Chapter 4's content is now complete.
I'd still like to swap the match statement at line 177 for a hash map, but I'd prefer not to make a new hash map for every single instance of LoxScanner, so I need to spend more time with the language to see if I can find a way to avoid that.